// src/app/Model/gradetype.ts
export class GradeType {
    companyId: string; 
      divisionId: string; 
      departmentId: string; 
      locationId:string 
      description: string;
      lockedBy: string;
      lockTs: Date;
      glaccountNumber: string;
      branchCode: any; 
      data:string;
    status: string;
  
  }   